Rename parser to language
